refactor: handleRegister in user/Handlers
