+++ src/lib/services/user/Handlers.ts
@@ -6,6 +6,9 @@ import { typeToFlattenedError } from 'zod'
 import { LoginRequestBodySchema } from '@/app/api/auth/login/route'
 import UserService from './UserService'
 import { isValidEmail } from '@/lib/util/util'
+import { RegisterRequestBodySchema } from '@/app/api/auth/register/route'
+import { UserRoleWithoutAdmin } from '@/types/User'
+import { validateEmail, validatePassword } from '@/utils/authUtils'
 
 /**
  * This function handles the form submission for user login.
@@ -56,5 +59,46 @@ export async function handleLoginButtonClick() {
     redirect('/')
   } else {
     redirect('/auth/login')
+  }
+}
+
+/**
+ * Handles the click event for register button
+ * Redirects to the appropriate page based on the login status.
+ */
+export const handleRegister = async (
+  formData: FormData,
+): Promise<void | {
+  error?: string
+  message?: string
+  details?: typeToFlattenedError<typeof RegisterRequestBodySchema>
+}> => {
+  const firstName = formData.get('firstName') as string
+  const lastName = formData.get('lastName') as string
+  const email = formData.get('email') as string
+  const password = formData.get('password') as string
+  const role = UserRoleWithoutAdmin.Client
+
+  if (!email) return { error: 'Email is required' }
+  if (!password) return { error: 'Password is required' }
+  if (!firstName) return { error: 'First name is required' }
+  if (!lastName) return { error: 'Last name is required' }
+  if (!validateEmail(email)) return { error: 'Invalid email address' }
+  if (!validatePassword(password).isValid) {
+    return { error: validatePassword(password).error }
+  }
+
+  const { message, status, error, details } = await UserService.register({
+    email,
+    password,
+    firstName,
+    lastName,
+    role,
+  })
+
+  if (status === 201) {
+    redirect('/login')
+  } else {
+    return { error, message, details }
   }
 }
